apps/api/pyproject.toml
[project]
name = "vacation-control-api"
version = "0.1.0"
description = "Backend API for vacation control"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.2.0",
  "alembic>=1.13.0",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "bcrypt>=4.1.2,<5",
  "email-validator>=2.2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
